frontend working - add payment component to reduce complex of draft detail component
