+++ donetop-main/frontend/src/app/component/draft/detail/detail.component.scss
@@ -188,21 +188,12 @@ div.container {
     > div.public {
       margin-left: auto;
 
-      > button {
-        background: #828ce8;
+      > app-payment {
+        display: inline-block;
+      }
 
-        &:last-child {
-          margin-right: 0;
-        }
-
-        &.print {
-          background: #4c4f6f;
-        }
-
-        &:disabled {
-          background: #565d9c;
-          cursor: default;
-        }
+      > button.print {
+        background: #4c4f6f;
       }
     }
   }
@@ -238,73 +229,5 @@ div.container {
 
   .hidden {
     visibility: hidden;
-  }
-}
-
-app-modal {
-  user-select: text;
-
-  h3.title {
-    height: 5%;
-    margin: 15px;
-    padding-left: 8px;
-  }
-
-  div.contents {
-    height: 80%;
-    margin: 0 15px;
-    display: flex;
-    flex-direction: column;
-
-    > div.row {
-      flex-grow: 1;
-      border-top: 1px solid #e6e6e6;
-      display: flex;
-
-      &:last-child {
-        border-bottom: 1px solid #e6e6e6;
-      }
-
-      > div {
-        padding-left: 8px;
-        display: flex;
-
-        &.key {
-          height: 100%;
-          width: 50%;
-          background: #f6f6f6;
-        }
-
-        &.value {
-          height: 100%;
-          width: 50%;
-        }
-
-        > div {
-          width: 100;
-          height: 100;
-          align-self: center;
-        }
-      }
-    }
-  }
-
-  div.buttons {
-    height: 15%;
-    margin: 15px;
-    text-align: center;
-
-    button {
-      line-height: 35px;
-      height: 35px;
-      padding: 0 10px;
-      text-align: center;
-      border-radius: 3px;
-      font-weight: bold;
-      border: 0;
-      margin-right: 5px;
-      color: #fff;
-      background: #4d5070;
-    }
   }
 }
